(PC-25369) refactor(env): env fixture in fixture file
