+++ src/features/profile/pages/PersonalData/PersonalData.native.test.tsx
@@ -6,7 +6,7 @@ import * as Auth from 'features/auth/context/AuthContext'
 import * as OpenUrlAPI from 'features/navigation/helpers/openUrl'
 import { beneficiaryUser } from 'fixtures/user'
 import { analytics } from 'libs/analytics'
-import { env } from 'libs/environment/__mocks__/envFixtures'
+import { env } from 'libs/environment/fixtures'
 import { mockServer } from 'tests/mswServer'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { render, fireEvent, screen, waitFor } from 'tests/utils'
