+++ src/libs/environment/__mocks__/useWebAppUrl.ts
@@ -1,3 +1,3 @@
-import { env } from './envFixtures'
+import { env } from '../fixtures'
 
 export const WEBAPP_V2_URL = `https://${env.WEBAPP_V2_DOMAIN}`
